BROOKLYN-532: karaf separate mgmt init and start
