+++ karaf/features/src/main/feature/feature.xml
@@ -341,6 +341,7 @@
         <feature>brooklyn-container-service</feature>
         <bundle dependency="true">mvn:org.apache.brooklyn/brooklyn-launcher-common/${project.version}</bundle>
         <bundle start-level="90">mvn:org.apache.brooklyn/brooklyn-karaf-init/${project.version}</bundle>
+        <bundle start-level="90">mvn:org.apache.brooklyn/brooklyn-karaf-start/${project.version}</bundle>
     </feature>
 
 </features>
