+++ karaf/pom.xml
@@ -61,6 +61,7 @@
 
   <modules>
     <module>init</module>
+    <module>start</module>
     <module>jetty-config</module>
     <module>httpcomponent-extension</module>
     <module>features</module>
